+++ src/components/ValidatedInput/ValidSelect.tsx
@@ -1,6 +1,6 @@
 /*
  *
- *  * Copyright (c) 2021 AccelByte Inc. All Rights Reserved.
+ *  * Copyright (c) 2021 - 2022 AccelByte Inc. All Rights Reserved.
  *  * This is licensed software from AccelByte Inc, for limitations
  *  * and restrictions contact your company contract manager.
  *
@@ -21,13 +21,9 @@ export interface ValidSelectProps extends SelectProps {
   options: SelectOption[];
   name?: string;
   value: SelectOption | SelectOption[] | string | null;
-  reloaderStatus?: Enum<typeof RELOADSTATUS>;
   errMessage?: string;
   onChange: (option: SelectOption | SelectOption[]) => void;
-  reloadMessage?: string;
-  onReload?: () => void;
   optionalLabel?: string;
-  isShowReloadWhileFetchFailed?: boolean;
   isInvalid?: boolean;
   isDisabled?: boolean;
   dataQa?: string | null;