+++ src/components/index.ts
@@ -63,3 +63,5 @@ export * from "./ValidatedInput/ValidFieldTextArea";
 export * from "./ValidatedInput/ValidMultiSelect";
 export * from "./ValidatedInput/ValidSelect";
 export * from "./ValidatedInput/ValidSelectAsync";
+export * from "./SelectAsync";
+export * from "./SelectWithLogo/SelectWithLogo"