+++ src/typings.d.ts
@@ -12,3 +12,5 @@ declare module "*.scss" {
   const classNames: IClassNames;
   export = classNames;
 }
+
+declare module 'simplebar-react'